style(static): Remove unnecessary u32 cast in sourcemap merge
Remove redundant type cast as sm.get_source_count() already returns u32.

Changes:
- crates/reinhardt-contrib/crates/static/src/processing/sourcemap.rs:368
  - Before: current_line += sm.get_source_count() as u32;
  - After:  current_line += sm.get_source_count();

This fixes clippy::unnecessary_cast warning in the SourceMapMerger::merge
method.

+++ crates/reinhardt-contrib/crates/static/src/processing/sourcemap.rs
@@ -365,7 +365,7 @@ impl SourceMapMerger {
 				}
 
 				// Update line offset for next map (approximate based on sources count)
-				current_line += sm.get_source_count() as u32;
+				current_line += sm.get_source_count();
 			}
 
 			// Collect sources and names for metadata
